Refactor YShrink JarStreamProvider to generic interface.
