+++ retroguard/src/main/java/com/yworks/util/abstractjar/Entry.java
@@ -4,7 +4,7 @@ package com.yworks.util.abstractjar;
  * Entry represents an entry in an archive, e.g class file or resource file
  */
 public interface Entry {
-  public boolean isDirectory();
-  public String getName();
-  public long getSize();
+  boolean isDirectory();
+  String getName();
+  long getSize();
 }
