+++ src/main/java/com/yworks/yshrink/core/Analyzer.java
@@ -11,9 +11,9 @@ import com.yworks.yshrink.model.Invocation;
 import com.yworks.yshrink.model.MethodDescriptor;
 import com.yworks.yshrink.model.Model;
 import com.yworks.yshrink.model.ModelVisitor;
-import com.yworks.yshrink.util.JarStreamProvider;
+import com.yworks.util.abstractjar.impl.JarStreamProvider;
 import com.yworks.yshrink.util.Logger;
-import com.yworks.yshrink.util.StreamProvider;
+import com.yworks.util.abstractjar.StreamProvider;
 import com.yworks.yshrink.util.Util;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassVisitor;
